+++ src/main/java/io/github/vikie1/portfolio/articles/ArticlesRepo.java
@@ -4,6 +4,6 @@ import java.util.List;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ArticlesRepo extends CrudRepository<Articles, Integer>{
+public interface ArticlesRepo extends CrudRepository<Articles, Long>{
     List<Articles> findAllByTopic(String topic);
 }
